fix the component
