+++ src/app/js/formats/streamgraph/Streamgraph.js
@@ -1,9 +1,8 @@
 import React, { PureComponent } from 'react';
 import { connect } from 'react-redux';
 import compose from 'recompose/compose';
-import differenceBy from 'lodash/differenceBy';
 
-import { draft } from './utils';
+import { getCssToString, zoomFunction, distinctColors, transformDataIntoMapArray, randomId, getMinMaxValue } from './utils';
 import injectData from '../injectData';
 import exportableToPng from '../exportableToPng';
 import * as d3 from 'd3';
@@ -36,226 +35,6 @@ const styles = {
     },
 };
 
-let xAxisScale;
-let yAxisScale;
-let xAxis;
-let yAxisL;
-let yAxisR;
-let gx;
-let gyr;
-let gyl;
-let streams;
-
-// ===========================================================================================
-// getCssToString get the css js object and return it as string
-// ===========================================================================================
-
-function getCssToString(cssToChange) {
-    const cssString = Object.keys(styles.legend).map(key => {
-        return `${key}: ${styles.legend[key]};`
-    }).join(" ");
-    return (cssString);
-}
-
-// ===========================================================================================
-// zoomFunction will update the zoom on the graph
-// ===========================================================================================
-
-function zoomFunction() {
-    // create new scale ojects based on event
-    let new_xScale = d3.event.transform.rescaleX(xAxisScale);
-    let new_yScale = d3.event.transform.rescaleY(yAxisScale);
-
-    // update axes
-    gx.call(xAxis.scale(new_xScale));
-    gyr.call(yAxisR.scale(new_yScale));
-    gyl.call(yAxisL.scale(new_yScale));
-
-    // update streams
-    streams.attr("transform", d3.event.transform);
-};
-
-// ===========================================================================================
-// distinctColors : return a list of @count different and distinct colors in hsl hexa
-// ===========================================================================================
-
-function distinctColors(count) {
-    let colors = [];
-    for (let hue = 0; hue < 360; hue += 360 / count) {
-        colors.push(hslToHex(hue, 90, 50));
-    }
-    return colors;
-}
-
-// ===========================================================================================
-// hslToHex : Convert hsl values color to hex color
-// ===========================================================================================
-
-function hslToHex(h, s, l) {
-    h /= 360;
-    s /= 100;
-    l /= 100;
-    let r, g, b;
-    if (s === 0) {
-        r = g = b = l; // achromatic
-    } else {
-        const hue2rgb = (p, q, t) => {
-            if (t < 0) t += 1;
-            if (t > 1) t -= 1;
-            if (t < 1 / 6) return p + (q - p) * 6 * t;
-            if (t < 1 / 2) return q;
-            if (t < 2 / 3) return p + (q - p) * (2 / 3 - t) * 6;
-            return p;
-        };
-        const q = l < 0.5 ? l * (1 + s) : l + s - l * s;
-        const p = 2 * l - q;
-        r = hue2rgb(p, q, h + 1 / 3);
-        g = hue2rgb(p, q, h);
-        b = hue2rgb(p, q, h - 1 / 3);
-    }
-    const toHex = x => {
-        const hex = Math.round(x * 255).toString(16);
-        return hex.length === 1 ? '0' + hex : hex;
-    };
-    return `#${toHex(r)}${toHex(g)}${toHex(b)}`;
-}
-
-// ===========================================================================================
-// transformDataIntoMapArray : convert all the data into an array
-// ===========================================================================================
-
-function transformDataIntoMapArray(formatData) {
-    console.log("========================================================");
-    console.log("formatData :");
-    console.log(formatData);
-    let dateMin = -42;
-    let dateMax = -42;
-    let valuesObjectsArray = [];
-
-    if (formatData) {
-        for (let i = 0; i < formatData.length; i++) {
-            let elem = formatData[i];
-            if (elem) {
-                // source: "2010"
-                // target: "Pétrogenèse de roches basaltiques"
-                // weight: 0.9013102637325718
-
-                if (elem.source && elem.weight) {
-
-                    let currentItem = undefined;
-
-                    for (let tmpElem of valuesObjectsArray) {
-                        if (tmpElem.name === elem.target) {
-                            currentItem = tmpElem;
-                        }
-                    }
-
-                    if (currentItem === undefined) {
-                        currentItem = {
-                            name: elem.target,
-                            values: []
-                        }
-                        valuesObjectsArray.push(currentItem);
-                    }
-
-                    currentItem.values.push({
-                        date: new Date(elem.source),
-                        value: elem.weight
-                    });
-
-                    // save min date and maxDate
-                    if (parseInt(elem.source) < dateMin || dateMin === -42) {
-                        dateMin = parseInt(elem.source);
-                    }
-                    if (parseInt(elem.source) > dateMax || dateMax === -42) {
-                        dateMax = parseInt(elem.source);
-                    }
-                }
-            }
-        }
-        // display the datas at the center when only one x value
-        if (dateMin == dateMax){
-            dateMin --;
-            dateMax ++;
-        }
-    }
-
-    let namesList = [];
-    for (let element of valuesObjectsArray) {
-        namesList.push(element.name)
-    }
-
-    let currentDate = dateMin;
-    let valuesArray = [];
-
-    while (currentDate <= dateMax) {
-        let tmpName = [];
-        let newElem = {
-            date: new Date(String(currentDate))
-        };
-
-        for (let element of valuesObjectsArray) {
-            // loop which add each values for the good year
-            for (let dateValue of element.values) {
-                if (dateValue.date.getFullYear() == currentDate) {
-                    newElem[element.name] = dateValue.value;
-                    tmpName.push(element.name);
-                }
-            }
-
-            // loop which add 0 value to the missing keys
-            let resMissingNameList = differenceBy(namesList, tmpName);
-            for (let elemToAdd of resMissingNameList) {
-                newElem[elemToAdd] = 0;
-            }
-        }
-
-        valuesArray.push(newElem);
-        currentDate++;
-    }
-
-    console.log("valuesObjectsArray");
-    console.log(valuesObjectsArray);
-    console.log("valuesArray");
-    console.log(valuesArray);
-    console.log("namesList");
-    console.log(namesList);
-    return ([valuesObjectsArray, valuesArray, dateMin, dateMax, namesList]);
-}
-
-// ===========================================================================================
-// getMinMaxValue : go through the stackedData and find the min and max value 
-// ===========================================================================================
-
-function getMinMaxValue(stackedData) {
-    let minValue = 0;
-    let maxValue = 0;
-
-    for (let element of stackedData) {
-        for (let value of element) {
-            if (minValue > value[0]) {
-                minValue = value[0];
-            }
-            if (maxValue < value[1]) {
-                maxValue = value[1];
-            }
-        }
-    }
-
-    return ([minValue, maxValue]);
-}
-
-const randomId = (length = 8) => {
-    let text = ''
-    const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
-
-    for (let i = 0; i < length; i++) {
-        text += chars.charAt(Math.floor(Math.random() * chars.length))
-    }
-
-    return text
-}
-
 class Streamgraph extends PureComponent {
 
     constructor(props) {
@@ -264,12 +43,24 @@ class Streamgraph extends PureComponent {
             width: 800,
             height: 300
         }
+
         this.uniqueId = randomId(10);
         this.divContainer = "divContainer" + this.uniqueId;
         this.svgContainer = "svgContainer" + this.uniqueId;
         this.anchor = "anchor" + this.uniqueId;
-        console.log("uniqueId : " + this.uniqueId);
-        this.draft = draft.bind(this);
+
+        this.xAxisScale;
+        this.yAxisScale;
+        this.xAxis;
+        this.yAxisL;
+        this.yAxisR;
+        this.gx;
+        this.gyr;
+        this.gyl;
+        this.streams;
+
+        //this.draft = draft.bind(this);
+        this.zoomFunction = zoomFunction.bind(this);
     }
 
     setGraph() {
@@ -282,9 +73,10 @@ class Streamgraph extends PureComponent {
 
         const d3DivContainerId = "d3DivContainer" + this.uniqueId;
         const divContainer = d3.select(this.refs.divContainer)
-        const d3DivContainer =  divContainer.append("div")
+        const d3DivContainer = divContainer.append("div")
             .attr("id", d3DivContainerId);
         const svgViewport = d3.select(this.refs.anchor);
+        let componentContext = this;
 
         // ===========================================================================================
         // Set all the variables (to move in states)
@@ -319,7 +111,7 @@ class Streamgraph extends PureComponent {
             .scaleExtent([1, 32])
             .translateExtent([[0, 0], [width, height]])
             .extent([[0, 0], [width, height]])
-            .on("zoom", zoomFunction);
+            .on("zoom", this.zoomFunction);
 
         // Inner Drawing Space
         const innerSpace = svgViewport.append("g")
@@ -355,36 +147,36 @@ class Streamgraph extends PureComponent {
         maxDate = new Date(String(maxDate));
 
         // create scale objects
-        xAxisScale = d3.scaleTime()
+        this.xAxisScale = d3.scaleTime()
             .domain([minDate, maxDate])
             .range([0, width]);
 
-        xAxis = d3.axisBottom(xAxisScale)
+        this.xAxis = d3.axisBottom(this.xAxisScale)
             .tickFormat(d3.timeFormat("%Y"))
             .ticks(d3.timeYear, 1);
 
-        gx = innerSpace.append("g")
+        this.gx = innerSpace.append("g")
             .attr("class", "x axis")
             .attr("transform", "translate(0," + height + ")")
-            .call(xAxis);
+            .call(this.xAxis);
 
-        yAxisScale = d3.scaleLinear()
+        this.yAxisScale = d3.scaleLinear()
             .domain([minValue, maxValue])
             .range([height, 0]);
 
-        yAxisL = d3.axisLeft(yAxisScale);
-        yAxisR = d3.axisRight(yAxisScale);
+        this.yAxisL = d3.axisLeft(this.yAxisScale);
+        this.yAxisR = d3.axisRight(this.yAxisScale);
 
         // append y axis to the right of the chart    
-        gyr = innerSpace.append("g")
+        this.gyr = innerSpace.append("g")
             .attr("class", "y axis")
             .attr("transform", "translate(" + width + ", 0)")
-            .call(yAxisR);
+            .call(this.yAxisR);
 
         // append y axis to the left of the chart
-        gyl = innerSpace.append("g")
+        this.gyl = innerSpace.append("g")
             .attr("class", "y axis")
-            .call(yAxisL);
+            .call(this.yAxisL);
 
         // color set
         const z = distinctColors(layersNumber);
@@ -396,30 +188,30 @@ class Streamgraph extends PureComponent {
         // ===========================================================================================
 
         const area = d3.area()
-            .x(function (d, i) {
-                return (xAxisScale(d.data.date));
+            .x((d, i) => {
+                return (this.xAxisScale(d.data.date));
             })
-            .y0(function (d) { return yAxisScale(d[0]); })
-            .y1(function (d) { return yAxisScale(d[1]); })
+            .y0((d) => { return this.yAxisScale(d[0]); })
+            .y1((d) => { return this.yAxisScale(d[1]); })
 
-        streams = graphZone.selectAll("path")
+        this.streams = graphZone.selectAll("path")
             .data(stackedData)
             .enter().append("path") // append all paths (area streams)
-            .attr("d", function (d) {
+            .attr("d", (d) => {
                 if (d) {
                     return area(d);
                 } else {
                     return ([]);
                 }
             })
-            .attr("name", function (d, i) {
+            .attr("name", (d, i) => {
                 let name = nameList[i];
                 let color = z[i];
                 colorNameList.push([name, color]);
                 return (name);
             })
             .attr("opacity", 1)
-            .attr("fill", function (d, i) { return (z[i]); });
+            .attr("fill", (d, i) => { return (z[i]); });
 
         // ===========================================================================================
         // Set tooltip
@@ -429,7 +221,7 @@ class Streamgraph extends PureComponent {
         const tooltip = d3DivContainer.append("div")
             .attr("class", "remove")
             .attr("id", "tooltip")
-            .attr("style", getCssToString(styles.tooltip))
+            .attr("style", getCssToString(styles.tooltip, styles))
             .style("position", "absolute")
             .style("z-index", "2")
             .style("visibility", "hidden")
@@ -439,7 +231,7 @@ class Streamgraph extends PureComponent {
         let hoveredKey;
         let hoveredValue;
 
-        // ===========================================================================================
+        // ==============================================================================differenceBy==========
         // Set vertical line
         // ===========================================================================================
 
@@ -465,12 +257,12 @@ class Streamgraph extends PureComponent {
 
         const legendView = d3DivContainer.append('div')
             .attr("id", "legend")
-            .attr("style", getCssToString(styles.legend))
+            .attr("style", getCssToString(styles.legend, styles))
 
         for (let index = colorNameList.length - 1; index > 0; index--) {
             let element = colorNameList[index];
             let legendItemContainer = legendView.append("div")
-                .attr("style", getCssToString(styles.legendItem))
+                .attr("style", getCssToString(styles.legendItem, styles))
                 .attr("class", "legendItem");
 
             legendItemContainer.append("svg")
@@ -479,7 +271,7 @@ class Streamgraph extends PureComponent {
                 .style("background-color", element[1]);
 
             legendItemContainer.append("text")
-                .attr("style", getCssToString(styles.legentItemText))
+                .attr("style", getCssToString(styles.legentItemText, styles))
                 .attr("class", "legentItemText")
                 .text(element[0]);
         }
@@ -524,7 +316,8 @@ class Streamgraph extends PureComponent {
         // Mouse events - streams events
         // ===========================================================================================
 
-        streams.on("mousemove", function (d, i) {
+
+        this.streams.on("mousemove", function (d, i) {
 
             // *******************************************************
             // Get the hovered values for tooltip
@@ -532,15 +325,9 @@ class Streamgraph extends PureComponent {
             // *******************************************************
 
             let mousex = d3.mouse(this)[0];
-            let date = xAxisScale.invert(mousex);
-            console.log("mousex : " + mousex);
-            console.log("date : " + date);
+            let date = componentContext.xAxisScale.invert(mousex);
 
             hoveredKey = d3.select(this).attr("name");
-            console.log("hoveredKey : " + hoveredKey);
-
-            console.log("range");
-            console.log(xAxisScale.range());
 
             for (let elem of d) {
                 if (elem.data.date.getFullYear() == date.getFullYear()) {
@@ -554,45 +341,44 @@ class Streamgraph extends PureComponent {
                 .attr("stroke", "#000")
                 .attr("stroke-width", "0.5px"),
                 tooltip.html("<p>" + hoveredKey + "<br>" + hoveredValue + "</p>").style("visibility", "visible");
-        }).on("mouseover", function (d, i) {
+        }).on("mouseover", (d, i) => {
 
             // *******************************************************
             // set the hover opacity
             // *******************************************************
 
             mouseIsOverStream = true;
-            streams.transition()
+            this.streams.transition()
                 .duration(25)
-                .attr("opacity", function (d, j) {
+                .attr("opacity", (d, j) => {
                     return j != i ? 0.3 : 1;
                 })
         }).on("mouseout", function (d, i) {
 
-                // *******************************************************
-                // unset hover and hide tooltip
-                // *******************************************************
+            // *******************************************************
+            // unset hover and hide tooltip
+            // *******************************************************
 
-                mouseIsOverStream = false;
-                streams.transition()
-                    .duration(25)
-                    .attr("opacity", "1");
-                d3.select(this)
-                    .classed("hover", false)
-                    .attr("stroke-width", "0px"),
-                    tooltip.html("<p>" + hoveredKey + "<br>" + hoveredValue + "</p>").style("visibility", "hidden");
-            });
+            mouseIsOverStream = false;
+            componentContext.streams.transition()
+                .duration(25)
+                .attr("opacity", "1");
+            d3.select(this)
+                .classed("hover", false)
+                .attr("stroke-width", "0px"),
+                tooltip.html("<p>" + hoveredKey + "<br>" + hoveredValue + "</p>").style("visibility", "hidden");
+        });
     }
 
     componentDidMount() {
         this.setGraph();
-        console.log("componentDIdMount");
     }
 
     componentWillUpdate() {
-        d3.selectAll("#d3DivContainer"+this.uniqueId).selectAll("div").remove();
-        d3.selectAll("#d3DivContainer"+this.uniqueId).remove();
-        d3.selectAll("#"+this.anchor).selectAll("g").remove();
-        d3.selectAll("#"+this.anchor).selectAll("defs").remove();
+        d3.selectAll("#d3DivContainer" + this.uniqueId).selectAll("div").remove();
+        d3.selectAll("#d3DivContainer" + this.uniqueId).remove();
+        d3.selectAll("#" + this.anchor).selectAll("g").remove();
+        d3.selectAll("#" + this.anchor).selectAll("defs").remove();
     }
 
     componentDidUpdate() {
@@ -612,8 +398,8 @@ class Streamgraph extends PureComponent {
 };
 
 const mapStateToProps = (state, { formatData }) => {
-    console.log("mapStateToProps - state:");
-    console.log(state);
+    //console.log("mapStateToProps - state:");
+    //console.log(state);
     return ({});
 };
